+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.3.0-beta.3
+
+- Fixed normal triggers restyling beyond the trigger string
+
 ## 2.3.0-beta.2
 
 - Fixed style string restyle being overwritten